+++ BL/Ibl.cs
@@ -26,7 +26,7 @@ namespace BL
         Child getChild(long idChild);
 
         void addContract(Contract contract);
-        void deleteContract(int idChildContractDel);
+        void deleteContract(long idChildContractDel);
         void updateContract(Contract contract);
         double getSalary(long idChild, long idNanny, bool isHour);
         Contract getContract(int idChildContract);
